requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.9
requests>=2.25
pandas>=1.2
numpy>=1.20
plotly>=4.14
exchange-calendars>=3.4
dateparser>=1.0
pytz>=2021.1
pytest>=6.2
pytest-django>=4.2
black>=21.5b2
isort>=5.8
flake8>=3.9
python-json-logger>=2.0.2
colorama>=0.4.4
gunicorn>=20.1.0
Sphinx>=4.0
python-dotenv>=0.17